refactor(gateway): make rustyclaw-gateway a bin-only crate
The gateway crate exists to provide the `rustyclaw-gateway` server binary, so
it should not also expose a library. Remove the lib target and fold the server
implementation into the binary's crate root (src/main.rs), so the crate has a
single binary target and no public API surface.

To break the last library dependency, `rustyclaw gateway run` (foreground) now
launches the gateway binary as a subprocess instead of calling the library
in-process — mirroring how `gateway start` already spawns it as a daemon. A new
`daemon::run_foreground` helper (sibling to `daemon::start`) resolves and runs
the binary with inherited stdio, blocking until exit. The interactive vault
prompt, model resolution, and skill loading all already live in the gateway
binary's main, so this is a faithful (and richer) replacement for the old
in-process path.

Knock-on changes:
- rustyclaw-cli drops its `rustyclaw-gateway` dependency entirely (it only
  locates the binary by name, via daemon::resolve_gateway_binary); the CLI
  `matrix` feature no longer forwards to the gateway crate.
- handle_run is now a thin sync wrapper that forwards --bind/--port plus TLS
  paths to the subprocess.
- The gateway crate's bin is discovered automatically from src/main.rs (no
  explicit [[bin]] / [lib]).

The daemon binary-resolution logic (look next to current exe, then $PATH) is
unchanged, so `gateway start`/`run` keep working — both binaries still build
into the same target dir.

Verified: cargo build --workspace (both binaries), --no-default-features,
-p rustyclaw-gateway --features all-messengers.
cargo clippy --workspace --all-targets -- -D warnings: clean.
cargo fmt --all -- --check: clean.
