+++ crates/rustyclaw-cli/Cargo.toml
@@ -17,13 +17,12 @@ workspace = true
 
 [features]
 default = ["tui", "desktop"]
-matrix = ["rustyclaw-core/matrix", "rustyclaw-gateway/matrix"]
+matrix = ["rustyclaw-core/matrix"]
 tui = ["dep:rustyclaw-tui"]
 desktop = ["dep:rustyclaw-desktop"]
 
 [dependencies]
 rustyclaw-core.workspace = true
-rustyclaw-gateway.workspace = true
 rustyclaw-onboard.workspace = true
 rustyclaw-tui = { workspace = true, optional = true }
 rustyclaw-desktop = { path = "../rustyclaw-desktop", optional = true }
@@ -47,7 +46,3 @@ libc = "0.2"
 [[bin]]
 name = "rustyclaw"
 path = "src/main.rs"
-
-[[bin]]
-name = "rustyclaw-gateway"
-path = "src/bin/rustyclaw-gateway.rs"
